chapter 2/requirements.txt
openai-agents==0.13.2
pytest==9.0.2
